refactor(orchestration): extract TaskLifecycleService
Twelve methods, 410 body lines, moved verbatim. 3113 -> 2693 lines. Nine
become one-line facade delegations; matchesFilter,
removeCoordinatorMetadataIfUnused and removeEmptyGroupsForCoordinator have no
caller outside the twelve and stay private on the service.

The plan's port list named closeWorkerSession and logger. Neither is used by
any of the twelve -- closeWorkerSession left the facade with
reconcileParallelSlots in Task 9, and every log goes through kernel.logEvent.
It also omitted config and createId, which listTasks and recordWorkerReply do
use. The Pick is now the five ports the code touches.

watchTask depends on five module-level free functions. The plan said to leave
them in orchestration-service.ts and import them from the service, but they
are values, not types: `verbatimModuleSyntax` would not erase that import, and
the facade imports the service class as a value, so it is a real runtime cycle.
Two constraints pin the fix -- the frozen oracle imports clampWatchTimeout from
orchestration-service and may not be touched, and nothing outside the facade
uses the other four. So clampWatchTimeout and clampWatchPollInterval move to
task-watch-timeouts.ts beside the constants they clamp; isTerminalTaskStatus
and isAttentionRequiredTask move to orchestration-types.ts as predicates over
the records defined there; sleep, which the plan's list omitted entirely, moves
into the service as a self-contained helper. orchestration-service.ts
re-exports clampWatchTimeout, so the oracle's import still resolves.

Four methods open kernel.mutate. None was added, removed, or relocated.
